Reworks Client class
